+++ src/app/app.component.spec.ts
@@ -24,7 +24,7 @@ describe('AppComponent', () => {
         expect(app).toBeTruthy();
     });
 
-    it('should have as title \'toolbox\'', () => {
+    it('should have as title \'ech-0160-dimag-ingest\'', () => {
         const fixture = TestBed.createComponent(AppComponent);
         const app = fixture.componentInstance;
         expect(app.navItems[0].label).toEqual('Playground');
@@ -34,6 +34,6 @@ describe('AppComponent', () => {
         const fixture = TestBed.createComponent(AppComponent);
         fixture.detectChanges();
         const compiled = fixture.nativeElement as HTMLElement;
-        expect(compiled.querySelector('h1.main-title')?.textContent).toContain('Toolbox für eCH-0160 Ingest Prozess');
+        expect(compiled.querySelector('h1.main-title')?.textContent).toContain('Applikation für eCH-0160 Ingest Prozess');
     });
 });